--- fe/fe-core/src/main/java/org/apache/doris/statistics/StatisticsDesc.java
@@ -1,61 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package org.apache.doris.statistics;
-
-import java.util.List;
-
-public class StatisticsDesc {
-    private StatsCategory statsCategory;
-
-    private StatsGranularity statsGranularity;
-
-    private List<StatsType> statsTypes;
-
-    public StatisticsDesc(StatsCategory statsCategory,
-                          StatsGranularity statsGranularity,
-                          List<StatsType> statsTypes) {
-        this.statsCategory = statsCategory;
-        this.statsGranularity = statsGranularity;
-        this.statsTypes = statsTypes;
-    }
-
-    public StatsCategory getStatsCategory() {
-        return statsCategory;
-    }
-
-    public void setStatsCategory(StatsCategory statsCategory) {
-        this.statsCategory = statsCategory;
-    }
-
-    public StatsGranularity getStatsGranularity() {
-        return statsGranularity;
-    }
-
-    public void setStatsGranularity(StatsGranularity statsGranularity) {
-        this.statsGranularity = statsGranularity;
-    }
-
-    public List<StatsType> getStatsTypes() {
-        return statsTypes;
-    }
-
-    public void setStatsTypes(List<StatsType> statsTypes) {
-        this.statsTypes = statsTypes;
-    }
-}
-
